Move Dealer Strategy Classes to Subdirectory
